refactor: 🔍  use prefered import/export syntax
